[arcane,materials] Déplace 'MeshEnvironmentVariableRef' dans 'arcane/core/materials'.
